+++ Widget.php
@@ -4,22 +4,24 @@ namespace koolreport\instant;
 
 class Widget
 {
-    static function create($widgetClass,$widgetParams,$reportSettings=array())
+    static function create($widgetClass,$widgetParams,$assets=true)
     {
+        
         $params = array_merge(array(
             "@widgetClass"=>$widgetClass,
             "@widgetParams"=>$widgetParams,
-            "@reportSettings"=>$reportSettings
+            "@assets"=>$assets
         ),$_GET,$_POST);
+        
         $tempReport = new TempReport($params);
         $tempReport->render();
     }
-    static function html($widgetClass,$widgetParams,$reportSettings=array())
+    static function html($widgetClass,$widgetParams,$assets=true)
     {
         $params = array_merge(array(
             "@widgetClass"=>$widgetClass,
             "@widgetParams"=>$widgetParams,
-            "@reportSettings"=>$reportSettings
+            "@assets"=>$assets
         ),$_GET,$_POST);
         $tempReport = new TempReport($params);
         return $tempReport->render(true);